Ruleta dividida en archivos, pero no se guarda posiciion
